CB-10699: Attempt to uninstall mobilespec on Android emulator before run
