+++ lib/util.js
@@ -24,11 +24,14 @@
 module.exports = (function () {
 
     var os = require("os");
+    var shelljs = require("shelljs");
 
     // constants
     var ESCAPE    = String.fromCharCode(27);
     var RED_COLOR = ESCAPE + "[31m";
     var NO_COLOR  = ESCAPE + "[m";
+    var DEVICE_ROW_PATTERN = /(emulator|device|host)/m;
+    var HEADING_LINE_PATTERN = /List of devices/m;
 
     function medicLog(message) {
         console.log(RED_COLOR + "[MEDIC LOG " + new Date().toUTCString() + "]" + NO_COLOR + " " + message);
@@ -54,6 +57,23 @@ module.exports = (function () {
         return Math.ceil(seconds / 60);
     }
 
+    function countAndroidDevices() {
+        var listCommand = "adb devices";
+
+        medicLog("running:");
+        medicLog("    " + listCommand);
+
+        var numDevices = 0;
+        var result = shelljs.exec(listCommand, {silent: false, async: false});
+        result.output.split('\n').forEach(function (line) {
+            if (!HEADING_LINE_PATTERN.test(line) && DEVICE_ROW_PATTERN.test(line)) {
+                numDevices += 1;
+            }
+        });
+
+        return numDevices;
+    }
+
     return {
 
         // constants
@@ -71,6 +91,7 @@ module.exports = (function () {
         isWindows: isWindows,
         medicLog:  medicLog,
         contains:  contains,
-        secToMin:  secToMin
+        secToMin:  secToMin,
+        countAndroidDevices: countAndroidDevices
     };
 }());
